xml/parsing/manakai/fragment-1.dat

#data
hoge
#document-fragment
p
#errors
#document
| "hoge"

#data
<foo>aa</foo>bar
#document-fragment
div
#errors
#document
| <foo>
|   "aa"
| "bar"

#data
<aa/>bb
#document-fragment
div
#errors
#document
| <aa>
| "bb"

#data
<foo xmlns=""/>
#document-fragment
div
#errors
#document
| <{} foo>
|   xmlns xmlns=""

#data
<foo xmlns:hoge=""/>
#document-fragment
div
#errors
#document
| <foo>
|   xmlns hoge=""

#data
<foo xml:lang="abc"/>
#document-fragment
div
#errors
#document
| <foo>
|   xml lang="abc"

#data
<hoge/>
#document-fragment
{} div
#errors
#document
| <{} hoge>

#data
<hoge>aaaa
#document-fragment
foo
#errors
1;11;in body:#eof;hoge;;m
#document
| <hoge>
|   "aaaa"

#data
<hoge/>aaaa</foo>bb
#document-fragment
foo
#errors
1;12;stray end tag;;foo;m
#document
| <hoge>
| "aaaabb"

#data
<hoge/>aaaa</>bb
#document-fragment
foo
#errors
1;12;stray end tag;;;m
1;12;empty end tag;;;m
#document
| <hoge>
| "aaaabb"
